+++ android/app/src/main/res/layout/item_store.xml
@@ -1,93 +1,92 @@
 <?xml version="1.0" encoding="utf-8"?>
 <androidx.cardview.widget.CardView xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="100dp"
-    android:layout_marginStart="2dp"
-    android:layout_marginEnd="2dp"
-    android:orientation="horizontal"
-    android:padding="10dp"
-    android:weightSum="10"
-    app:cardCornerRadius="5dp"
-    app:cardElevation="5dp"
-    app:paddingEnd ="5dp">
+                                   xmlns:app="http://schemas.android.com/apk/res-auto"
+                                   xmlns:tools="http://schemas.android.com/tools"
+                                   android:layout_width="match_parent"
+                                   android:layout_height="100dp"
+                                   android:layout_marginStart="2dp"
+                                   android:layout_marginEnd="2dp"
+                                   android:orientation="horizontal"
+                                   android:padding="10dp"
+                                   android:weightSum="10"
+                                   app:cardCornerRadius="5dp"
+                                   app:cardElevation="5dp"
+                                   app:paddingEnd ="5dp">
 
     <LinearLayout
-        android:layout_width="match_parent"
-        android:layout_height="match_parent"
-        android:orientation="horizontal">
-        <LinearLayout
-            android:layout_width="0dp"
+            android:layout_width="match_parent"
             android:layout_height="match_parent"
-            android:layout_weight="2.5"
-            android:layout_marginRight="5dp"
-            android:gravity="center">
+            android:orientation="horizontal">
+        <LinearLayout
+                android:layout_width="0dp"
+                android:layout_height="match_parent"
+                android:layout_weight="2.5"
+                android:layout_marginRight="5dp"
+                android:gravity="center">
             <ImageView
-                android:id="@+id/item_image"
-                android:layout_width="80dp"
-                android:layout_height="80dp"
-                android:padding="10dp"
-                android:background="@drawable/gradient_green" />
+                    android:id="@+id/item_image"
+                    android:layout_width="80dp"
+                    android:layout_height="80dp"
+            />
         </LinearLayout>
 
 
         <LinearLayout
-            android:layout_width="0dp"
-            android:layout_height="match_parent"
-            android:layout_weight="7.5"
-            android:orientation="vertical"
-            android:weightSum="10">
-
-            <TextView
-                android:id="@+id/item_name"
-                android:layout_width="match_parent"
-                android:layout_height="0dp"
-                android:layout_weight="5"
-                android:fontFamily="@font/sc_6"
-                android:gravity="center|left"
-                android:text="이자카야"
-                android:textColor="#000000"
-                android:textSize="18dp" />
-
-            <LinearLayout
-                android:layout_width="match_parent"
-                android:layout_height="0dp"
-                android:layout_weight="2.5"
+                android:layout_width="0dp"
+                android:layout_height="match_parent"
+                android:layout_weight="7.5"
+                android:orientation="vertical"
                 android:weightSum="10">
 
+            <TextView
+                    android:id="@+id/item_name"
+                    android:layout_width="match_parent"
+                    android:layout_height="0dp"
+                    android:layout_weight="5"
+                    android:fontFamily="@font/sc_6"
+                    android:gravity="center|left"
+                    android:text="이자카야"
+                    android:textColor="#000000"
+                    android:textSize="18dp" />
+
+            <LinearLayout
+                    android:layout_width="match_parent"
+                    android:layout_height="0dp"
+                    android:layout_weight="2.5"
+                    android:weightSum="10">
+
                 <TextView
-                    android:id="@+id/item_address"
-                    android:layout_width="0dp"
-                    android:layout_height="match_parent"
-                    android:layout_weight="8.5"
-                    android:fontFamily="@font/sc_4"
-                    android:gravity="left|center"
-                    android:text="지구상어디인가"
-                    android:textSize="11dp" />
+                        android:id="@+id/item_address"
+                        android:layout_width="0dp"
+                        android:layout_height="match_parent"
+                        android:layout_weight="8.5"
+                        android:fontFamily="@font/sc_4"
+                        android:gravity="left|center"
+                        android:text="지구상어디인가"
+                        android:textSize="11dp" />
 
                 <Button
-                    android:id="@+id/item_copy"
-                    android:layout_width="0dp"
-                    android:layout_height="match_parent"
-                    android:layout_marginTop="1dp"
-                    android:layout_marginBottom="1dp"
-                    android:layout_marginRight="5dp"
-                    android:layout_weight="1.5"
-                    android:background="@drawable/btn_grayline"
-                    android:fontFamily="@font/sc_4"
-                    android:text="복사"
-                    android:textSize="8dp" />
+                        android:id="@+id/item_copy"
+                        android:layout_width="0dp"
+                        android:layout_height="match_parent"
+                        android:layout_marginTop="1dp"
+                        android:layout_marginBottom="1dp"
+                        android:layout_marginRight="5dp"
+                        android:layout_weight="1.5"
+                        android:background="@drawable/btn_grayline"
+                        android:fontFamily="@font/sc_4"
+                        android:text="복사"
+                        android:textSize="8dp" />
             </LinearLayout>
 
             <TextView
-                android:id="@+id/item_call"
-                android:layout_width="match_parent"
-                android:layout_height="0dp"
-                android:layout_weight="2.5"
-                android:fontFamily="@font/sc_4"
-                android:text="누군가의 번호 "
-                android:textSize="11dp" />
+                    android:id="@+id/item_call"
+                    android:layout_width="match_parent"
+                    android:layout_height="0dp"
+                    android:layout_weight="2.5"
+                    android:fontFamily="@font/sc_4"
+                    android:text="누군가의 번호 "
+                    android:textSize="11dp" />
 
 
         </LinearLayout>
